split index.js into two files - index.js that has the matching functionality and generatorsupport.js that has functionality needed to support generator.js
